src/requirements.txt
Cython==0.23.4
Flask==0.10.1
Jinja2==2.8
Lasagne==0.1
MarkupSafe==0.23
Pillow==3.1.0
Theano==0.7.0
Werkzeug==0.11.3
#h5py==2.5.0
itsdangerous==0.24
numpy>=1.16.0
scikit-learn==0.17
scipy==0.16.1
six==1.10.0
sklearn==0.0
wget==3.2
wsgiref==0.1.2
python-dotenv==0.13.0